# docs/backend/api.yaml
openapi: 3.0.3
info:
  title: Paperflow Java Admin API
  version: 0.1.0
  description: REST API over the Paperflow PostgreSQL database. Paperflow business data is read-only; Admin User management writes only admin_user.
servers:
  - url: http://localhost:8080
security:
  - sessionCookie: []
...
paths:
...
  /api/sources/{sourceId}:
    get:
      tags: [Sources]
      operationId: getSource
      summary: Get one Source with aggregate processing statistics.
      parameters:
        - $ref: "#/components/parameters/SourceId"
      responses:
        "200":
          description: Source detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceDetail"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          description: Source not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                sourceNotFound:
                  value:
                    code: SOURCE_NOT_FOUND
                    message: Source not found
                    requestId: 9c0d2d7e3f1a4b61
        "500":
          $ref: "#/components/responses/InternalError"
...
  /api/openalex/source-search:
    get:
      tags: [OpenAlex Sources]
      operationId: searchOpenAlexSources
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            maxLength: 200
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 20
      responses:
        "200":
          description: Source recommendations.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OpenAlexSource"
  /api/openalex/source-search/{sourceId}:
    get:
      tags: [OpenAlex Sources]
      operationId: getOpenAlexSource
      parameters:
        - $ref: "#/components/parameters/SourceId"
      responses:
        "200":
          description: Authoritative OpenAlex Source.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAlexSource"
...
  /api/assets/{relativePath}:
    get:
      tags: [Assets]
      operationId: getAsset
      summary: Read one registered file asset under DATA_ROOT.
      description: Streams a read-only file under DATA_ROOT. Clients receive asset URLs from block, text file, or Original File responses.
      parameters:
        - name: relativePath
          in: path
          required: true
          schema:
            type: string
          description: DATA_ROOT-relative path. May include slashes.
      responses:
        "200":
          description: File content.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Asset not found.
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
  parameters:
...
    SourceId:
      name: sourceId
      in: path
      required: true
      schema:
        type: string
        pattern: "^S.+"
...
  responses:
    ValidationError:
      description: Invalid request parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            validation:
              value:
                code: VALIDATION_ERROR
                message: Invalid request
                requestId: 9c0d2d7e3f1a4b61
...
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            internal:
              value:
                code: INTERNAL_ERROR
                message: Internal server error
                requestId: 9c0d2d7e3f1a4b61
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, requestId]
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - UNAUTHORIZED
            - FORBIDDEN
            - ADMIN_USER_NOT_FOUND
            - ADMIN_USER_CONFLICT
            - SOURCE_NOT_FOUND
            - WORK_NOT_FOUND
            - ORIGINAL_FILE_NOT_FOUND
            - PAPER_NOT_FOUND
            - PAPER_ALREADY_EXISTS
            - PAPER_IN_TRASH
            - PAPER_VERSION_CONFLICT
            - PAPER_MATCHED_FIELDS_LOCKED
            - PAPER_STORAGE_CONFLICT
            - INVALID_PAPER_FILE
            - PAPER_FILE_TOO_LARGE
            - PURGE_CONFIRMATION_REQUIRED
            - INTERNAL_ERROR
        message:
          type: string
        requestId:
          type: string
...
    SourceSummary:
      type: object
      required: [sourceId, stats]
      properties:
        sourceId:
          type: string
        sourceName:
          type: string
          nullable: true
        provider:
          type: string
          nullable: true
        stats:
          $ref: "#/components/schemas/SourceStats"
    SourceDetail:
      allOf:
        - $ref: "#/components/schemas/SourceSummary"
    SourceStats:
      type: object
      required:
        - workCount
        - originalFileCount
        - matchedFileCount
        - parsedFileCount
        - readyFileCount
        - parseFailedFileCount
        - blockFailedFileCount
        - unsupportedFileCount
      properties:
        workCount:
          type: integer
          format: int64
        originalFileCount:
          type: integer
          format: int64
        matchedFileCount:
          type: integer
          format: int64
        parsedFileCount:
          type: integer
          format: int64
        readyFileCount:
          type: integer
          format: int64
        parseFailedFileCount:
          type: integer
          format: int64
        blockFailedFileCount:
          type: integer
          format: int64
        unsupportedFileCount:
          type: integer
          format: int64
...
    OpenAlexSource:
      type: object
      required: [sourceId, displayName, issn]
      properties:
        sourceId:
          type: string
        displayName:
          type: string
        publisher:
          type: string
          nullable: true
        issnL:
          type: string
          nullable: true
        issn:
          type: array
          items:
            type: string
        worksCount:
          type: integer
          nullable: true
        citedByCount:
          type: integer
          nullable: true
        isOa:
          type: boolean
          nullable: true
        isInDoaj:
          type: boolean
          nullable: true
        homepageUrl:
          type: string
          nullable: true
...